+++ src/main/java/com/debuggeando_ideas/best_travel/domain/entities/ReservationEntity.java
@@ -7,6 +7,8 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
+import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.UUID;
@@ -32,7 +34,7 @@ public class ReservationEntity {
     private LocalDate dateStart;
     private LocalDate dateEnd;
     private Integer totalDays;
-    private Double price;
+    private BigDecimal price;
     @ManyToOne
     @JoinColumn(name = "hotel_id", referencedColumnName = "id")
     private HotelEntity hotel;
